API Move samplers/__init__ to samplers/base.py
